+++ src/main/java/gusta/miithersz/geekcave/repositories/anime/AnimeCharacterRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import gusta.miithersz.geekcave.models.anime.AnimeCharacterModel;
 
 @Repository
-public interface AnimeCharacterRepository extends JpaRepository<AnimeCharacterModel, Integer> {
+public interface AnimeCharacterRepository extends JpaRepository<AnimeCharacterModel, Long> {
     
 }
