+++ src/main/java/gusta/miithersz/geekcave/repositories/anime/AnimeStudioRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import gusta.miithersz.geekcave.models.anime.AnimeStudioModel;
 
 @Repository
-public interface AnimeStudioRepository extends JpaRepository<AnimeStudioModel, Integer> {
+public interface AnimeStudioRepository extends JpaRepository<AnimeStudioModel, Long> {
 
 }
