+++ src/main/java/gusta/miithersz/geekcave/repositories/anime/AnimeTitleRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import gusta.miithersz.geekcave.models.anime.AnimeTitleModel;
 
 @Repository
-public interface AnimeTitleRepository extends JpaRepository<AnimeTitleModel, Integer> {
+public interface AnimeTitleRepository extends JpaRepository<AnimeTitleModel, Long> {
 
 }
